requirements.txt
pandas==1.5.2
Pillow==9.3.0
requests==2.28.1
streamlit==1.15.1
streamlit_lottie==0.0.3
